+++ common/src/main/java/com/mrpgc_skill_tree/MRPGCSkillTreeAddOn.java
@@ -5,10 +5,7 @@ import com.mrpgc_skill_tree.effect.MrpgSkillEffects;
 import com.mrpgc_skill_tree.skills.CustomSpellImpacts;
 import com.mrpgc_skill_tree.skills.MrpgSkillSounds;
 import com.mrpgc_skill_tree.config.TweaksConfig;
-import net.fabricmc.fabric.api.resource.ResourceManagerHelper;
-import net.fabricmc.fabric.api.resource.ResourcePackActivationType;
 import net.fabricmc.loader.api.FabricLoader;
-import net.minecraft.util.Identifier;
 import net.spell_engine.api.config.ConfigFile;
 import net.tiny_config.ConfigManager;
 import org.slf4j.Logger;
@@ -24,7 +21,7 @@ public class MRPGCSkillTreeAddOn{
 			.sanitize(true)
 			.build();
 	public static ConfigManager<TweaksConfig> tweaksConfig = new ConfigManager<TweaksConfig>
-			("tweaks", new TweaksConfig())
+			("tweaks_v0", new TweaksConfig())
 			.builder()
 			.setDirectory(MOD_ID)
 			.sanitize(true)
@@ -39,18 +36,6 @@ public class MRPGCSkillTreeAddOn{
 		}
 		CustomSpellImpacts.registerCustomImpacts();
 
-	}
-	public static void registerItems() {
-		if (!MRPGCSkillTreeAddOn.tweaksConfig.value.disable_mrpgc_skilltree_changes) {
-			FabricLoader.getInstance().getModContainer(MOD_ID).ifPresent(modContainer -> {
-				ResourceManagerHelper.registerBuiltinResourcePack(
-						Identifier.of(MOD_ID, "mrpgc_skill_tree_changes"),
-						modContainer,
-						ResourcePackActivationType.ALWAYS_ENABLED
-				);
-			});
-		}
-		tweaksConfig.save();
 	}
 	public static void registerSounds() {
 		MrpgSkillSounds.register();